Add option to toggle option page visibility
There are situations in which it is desirable to register an option
page and its fields but omit the page from the admin menu. In these
situations, a developer can set `show_in_menu` to `false`, which
causes the instantiated options object to skip the call it would
otherwise make to `acf_add_options_page`.

This feature does not affect the existing expected functionality of
the plugin, as `show_in_menu` is by default set to `true`.